refactor: 타입 수정
Money에 더 큰 값이 들어올것을 대비하여 int -> long으로 수정

+++ src/main/java/lotto/controller/LottoController.java
@@ -47,7 +47,7 @@ public class LottoController {
     }
 
     private Lottos buyLottos(final Money money, final NumberGenerator numberGenerator) {
-        int buyCount = money.buyLotto();
+        long buyCount = money.buyLotto();
         outputView.printBuyLotto(buyCount);
         Lottos lottos = Lottos.of(buyCount, numberGenerator);
         printLottos(lottos);
